feat(collections): stabilize iterable input of takeWhile
