+++ collections/take_while_test.ts
@@ -2,7 +2,6 @@
 
 import { assertEquals } from "@std/assert";
 import { takeWhile } from "./take_while.ts";
-import { takeWhile as unstableTakeWhile } from "./unstable_take_while.ts";
 
 Deno.test("takeWhile() handles num array", () => {
   const arr = [1, 2, 3, 4, 5, 6];
@@ -57,60 +56,7 @@ Deno.test("takeWhile() returns the same array when all elements match the predic
   assertEquals(actual, [1, 2, 3, 4]);
 });
 
-Deno.test("(unstable) takeWhile() handles num array", () => {
-  const arr = [1, 2, 3, 4, 5, 6];
-  const actual = unstableTakeWhile(arr, (i) => i !== 4);
-
-  assertEquals(actual, [1, 2, 3]);
-});
-
-Deno.test("(unstable) takeWhile() adds two to each num in predicate", () => {
-  const arr = [1, 2, 3, 4, 5, 6];
-  const actual = unstableTakeWhile(arr, (i) => i + 2 !== 6);
-
-  assertEquals(actual, [1, 2, 3]);
-});
-
-Deno.test("(unstable) takeWhile() handles negatives", () => {
-  const arr = [-1, -2, -3, -4, -5, -6];
-
-  const actual = unstableTakeWhile(arr, (i) => i > -4);
-  assertEquals(actual, [-1, -2, -3]);
-});
-
-Deno.test("(unstable) takeWhile() handles no mutation", () => {
-  const arr = [-1, -2, -3, -4, -5, -6];
-
-  const actual = unstableTakeWhile(arr, (i) => i > -4);
-  assertEquals(actual, [-1, -2, -3]);
-  assertEquals(arr, [-1, -2, -3, -4, -5, -6]);
-});
-
-Deno.test("(unstable) takeWhile() handles empty input array returns empty array", () => {
-  const arr: number[] = [];
-
-  const actual = unstableTakeWhile(arr, (i) => i > 4);
-
-  assertEquals(actual, []);
-});
-
-Deno.test("(unstable) takeWhile() returns empty array when the first element doesn't match the predicate", () => {
-  const arr = [1, 2, 3, 4];
-
-  const actual = unstableTakeWhile(arr, (i) => i !== 1);
-
-  assertEquals(actual, []);
-});
-
-Deno.test("(unstable) takeWhile() returns the same array when all elements match the predicate", () => {
-  const arr = [1, 2, 3, 4];
-
-  const actual = unstableTakeWhile(arr, (i) => i !== 400);
-
-  assertEquals(actual, [1, 2, 3, 4]);
-});
-
-Deno.test("(unstable) takeWhile() handles a generator", () => {
+Deno.test("takeWhile() handles a generator", () => {
   function* infiniteCount() {
     let count = 0;
     while (true) {
@@ -118,17 +64,17 @@ Deno.test("(unstable) takeWhile() handles a generator", () => {
       yield count;
     }
   }
-  const actual = unstableTakeWhile(infiniteCount(), (i) => i !== 4);
+  const actual = takeWhile(infiniteCount(), (i) => i !== 4);
   assertEquals(actual, [1, 2, 3]);
 });
 
-Deno.test("(unstable) takeWhile() handles a Set", () => {
+Deno.test("takeWhile() handles a Set", () => {
   const set = new Set([1, 2, 3, 4, 5, 6]);
-  const actual = unstableTakeWhile(set, (i) => i !== 4);
+  const actual = takeWhile(set, (i) => i !== 4);
   assertEquals(actual, [1, 2, 3]);
 });
 
-Deno.test("(unstable) takeWhile() handles a Map", () => {
+Deno.test("takeWhile() handles a Map", () => {
   const map = new Map([
     ["a", 1],
     ["b", 2],
@@ -137,7 +83,7 @@ Deno.test("(unstable) takeWhile() handles a Map", () => {
     ["e", 5],
     ["f", 6],
   ]);
-  const actual = unstableTakeWhile(map, ([, i]) => i !== 4);
+  const actual = takeWhile(map, ([, i]) => i !== 4);
   assertEquals(actual, [
     ["a", 1],
     ["b", 2],
